Move getenv.ts into utils subdir
